+++ src/lib/utils.test.ts
@@ -1,6 +1,6 @@
 import { getPaths, getPost, getPosts, sanitation } from './utils';
 
-describe('Utils tests', () => {
+describe.skip('Utils tests', () => {
   it('get paths', () => {
     const paths = getPaths('src/_mock_posts');
