+++ src/pages/blog/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { GetStaticProps, NextPage } from 'next';
+import type { NextPage } from 'next';
 import Head from 'next/head';
 
 const Blog: NextPage = () => {
